# packages/db/src/schema/swaprequests.ts
import { pgTable, serial, text, timestamp, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod";
import { usersTable } from "./users";
import { skillPostsTable } from "./skillposts";

export const swapRequestsTable = pgTable("swap_requests", {
  id: serial("id").primaryKey(),
  senderId: integer("sender_id")
    .notNull()
    .references(() => usersTable.id, { onDelete: "cascade" }),
  receiverId: integer("receiver_id")
    .notNull()
    .references(() => usersTable.id, { onDelete: "cascade" }),
  skillPostId: integer("skill_post_id")
    .notNull()
    .references(() => skillPostsTable.id, { onDelete: "cascade" }),
  message: text("message"),
  status: text("status").notNull().default("pending"),
  createdAt: timestamp("created_at").defaultNow().notNull(),
});

export const insertSwapRequestSchema = createInsertSchema(swapRequestsTable).omit({
  id: true,
  createdAt: true,
  status: true,
});

export type InsertSwapRequest = z.infer<typeof insertSwapRequestSchema>;
export type SwapRequest = typeof swapRequestsTable.$inferSelect;